Make chain call function accessible for users
