+++ python3-job-type/plugin-manifest.yaml
@@ -1,3 +1,3 @@
 name: python3-job-type
-version: 2.7.0
+version: 2.8.0
 url: 'https://github.com/TheRacetrack/plugin-python-job-type'
