+++ python3-job-type/python_wrapper/requirements-dev.txt
@@ -4,4 +4,3 @@ coverage==7.2.1
 backoff==2.2.1
 numpy==1.24.2
 Flask==2.2.5
-httpx==0.23.3
