+++ python3-job-type/python_wrapper/requirements.txt
@@ -1,2 +1,3 @@
 PyYAML==6.0
 a2wsgi==1.7.0
+httpx==0.24.0
